Добавил базовую поддержку фильтрации событий аудита в REST endpoint

+++ src/main/java/org/samearch/jira/lib/entity/mapper/AuditJournalFilter.java
@@ -3,7 +3,7 @@ package org.samearch.jira.lib.entity.mapper;
 import java.util.Set;
 
 public interface AuditJournalFilter {
-    Long eventsCount();
+    Integer eventsCount();
     Set<Long> forIds();
     Set<String> byInitiator();
     DateRange inDateRange();
